Refactor: Split chat flow engine into modules
Split the chat flow engine into smaller, more manageable modules to improve code organization and readability.
