+++ src/utils/chat/index.ts
@@ -1,0 +1,4 @@
+
+// Central export file for chat utilities
+export { processConversation, applyTrinidadianStyle } from './chatFlowEngine';
+export { ChatConfig, defaultChatConfig } from './engine/types';
